+++ ZucchiniMVC/Services/API/IWeatherService.cs
@@ -6,9 +6,7 @@ namespace Zucchinimvc.Services.API
     public interface IWeatherService 
     {
         Task<WeatherViewModel?> GetWeatherByCityAsync(string city);
-
         Task<List<WeatherHistoryEntity>> GetWeatherAsync();
-
         Task<List<WeatherHistoryEntity>> GetHistoryAsync(string city);
 
     }
